pdf-pal/requirement.txt
streamlit
PyPDF2
langchain
langchain-google-genai
faiss-cpu
python-dotenv
google-generativeai













